backend/requirements.txt
omniclaw==0.0.7
fastapi==0.136.1
uvicorn==0.46.0
python-dotenv==1.2.2
httpx==0.28.1
pydantic==2.12.0
web3==6.11.0
setuptools==80.10.2
